Add files via upload
Loading the Data: The project begins by loading the provided dataset, which contains monthly milk production data over a period of time.
Exploratory Data Analysis (EDA): The dataset is explored to understand its structure, distribution, trends, and seasonality. Visualizations such as line plots and summary statistics are used to gain insights into the data.
Checking for Stationarity: Stationarity is a key assumption in time series analysis. The Augmented Dickey-Fuller (ADF) test is used to check for stationarity. If the data is not stationary, suitable methods like differencing are applied to make it stationary.
Building Forecasting Models: Various time series forecasting models are explored and implemented. In this project, an Autoregressive Integrated Moving Average (ARIMA) model is used. The ARIMA model parameters are determined based on the characteristics of the data and the results of the ACF (Autocorrelation Function) and PACF (Partial Autocorrelation Function) plots.
Model Evaluation and Prediction: The ARIMA model is trained on the dataset, and its performance is evaluated using appropriate metrics. The trained model is then used to make predictions for future periods, providing insights into potential future milk production trends.
Overall, this project aims to demonstrate the importance of accurate forecasting in the dairy industry and explores the use of advanced statistical and machine learning techniques to achieve this goal. By accurately forecasting milk production, stakeholders can optimize production processes, reduce wastage, and ensure a stable supply of milk in the market